Overall status: it should work. But it doesn't. Don't ask why, I don't know, some error about 0x200 not being a valid class attribute (it isn't, but it also isn't in the binary file...)
Bsp.java: modified to be closer to my example
ClassFormat.hs: minor change
Codegen.hs: bugfixes
Bytecodegen.hs: writes JavaBytecode
Constants.hs: added String reference type
Typecheck.hs: added casting of types to reference types (java/lang/Object <--> Ljava/lang/Object; etc.)

+++ ClassFormat.hs
@@ -76,7 +76,7 @@ data CP_Info =
                 }
         | Float_Info 
                 { tag_cp                :: Tag
-                , numf_cp               :: Float
+                , numf_cp               :: Int
                 , desc                  :: String
                 }
         | Long_Info
